test: session을 얻는 편의 메서드 등록
